Moved <MobileBookingHeader /> under src/MobileBookingHeader
